requirements.txt
pymongo
openai
# gemini-python
streamlit
google-genai
google-generativeai
langchain
langchain_community
pypdf
langchain-openai
langchain-google-genai